Allow half days selection
Change the working days field to a decimal and change all the relevant references.  Add in a new set of calculations to the excel document to help aid testing.

Update the About page to include details on the new version changes + changes that wern't documented for the patch release in the summer.

+++ Retirement Countdown Clock Core/uk/co/vsf/retirement/repository/WorkingDaysParametersRepositoryImpl.cs
@@ -33,7 +33,7 @@ namespace Retirement_Countdown_Clock_Core.uk.co.vsf.retirement.repository
         public WorkingDaysParameters RetrieveWorkingDaysParameters()
         {
             Windows.Storage.ApplicationDataContainer localSettings = GetLocalSettings();
-            int? workingDays = (int?)localSettings.Values[WD];
+            double? workingDays = (double?)localSettings.Values[WD];
             int? holidays = (int?)localSettings.Values[H];
             int? bankHolidays = (int?)localSettings.Values[BH];
